move marking algorithms and parts references to their own pages
The stuff on marking algorithms is very long, and not specific to custom
part types, so I've moved it to its own page.

I've documented the marking interface in the custom part type editor,
and the marking algorithm tab in the part editor.

I *think* I've now documented all the new stuff.
Needs screencasts and images about marking algorithms.